Refactoring Entity Relations to ManyToMany

+++ src/main/java/com/solo83/weatherapp/dto/LocationFromRequest.java
@@ -2,19 +2,17 @@ package com.solo83.weatherapp.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class LocationFromRequest {
     private Integer id;
     private String name;
     private String country;
     private String state;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private String temperature;
 
-    public LocationFromRequest(String name, BigDecimal latitude, BigDecimal longitude) {
+    public LocationFromRequest(String name, Double latitude, Double longitude) {
         this.name = name;
         this.latitude = latitude;
         this.longitude = longitude;
